pyproject.toml
[tool.poetry]
name = "tags-by-review-nlp"
version = "0.1.0"
description = ""
authors = ["Dmitry Nesmeyanov <miradg99@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "~3.10"


[tool.poetry.group.data.dependencies]
sqlalchemy = "^1.4"
pandas = "^2.1.1"
pyarrow = "^13.0.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.linters.dependencies]
black = "23.9.1"
isort = "5.12.0"
autopep8 = "2.0.4"
add-trailing-comma = "3.1.0"
pre-commit = "3.4.0"


[tool.poetry.group.support.dependencies]
pyyaml = "^6.0.1"
poetry-dotenv-plugin = "^0.2.0"
click = "^8.1.7"
dvc = {extras = ["s3"], version = "^3.23.0"}



[tool.poetry.group.models.dependencies]
transformers = {extras = ["sentencepiece"], version = "^4.33.3"}
torch = {url="https://download.pytorch.org/whl/cu117_pypi_cudnn/torch-2.0.1%2Bcu117.with.pypi.cudnn-cp310-cp310-linux_x86_64.whl"}
nltk = {extras = ["stopwords"], version = "^3.8.1"}
spacy = "^3.7.0"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
